requirements.txt
Flask
opencv-python
python-dotenv
anthropic
ultralytics
Pillow
torch
torchvision
gunicorn
